+++ packages/core/src/connectivity/scp-cf/token-accessor.ts
@@ -1,6 +1,5 @@
 import { ErrorWithCause } from '@sap-cloud-sdk/util';
 import { JwtPayload } from 'jsonwebtoken';
-import * as xssec from '@sap/xssec';
 import { decodeJwt } from './jwt';
 import { CachingOptions } from './cache';
 import { clientCredentialsTokenCache } from './client-credentials-token-cache';
@@ -11,46 +10,10 @@ import {
 } from './environment-accessor';
 import { Service } from './environment-accessor-types';
 import { ResilienceOptions } from './resilience-options';
-import { parseSubdomain, replaceSubdomain } from './subdomain-replacer';
-import {
-  jwtBearerTokenGrant,
-  refreshTokenGrant,
-  userTokenGrant
-} from './xsuaa-service';
-import {
-  ClientCredentialsResponse,
-  UserTokenResponse
-} from './xsuaa-service-types';
-
-async function getClientCredentialsToken(
-  service: string | Service,
-  userJwt?: string | JwtPayload
-): Promise<ClientCredentialsResponse> {
-  const serviceCredentials = resolveService(service).credentials;
-
-  let subdomain: string;
-  let zoneId: string;
-
-  if (userJwt) {
-    const jwtPayload =
-      typeof userJwt === 'string' ? decodeJwt(userJwt) : userJwt;
-    if (jwtPayload.iss) {
-      subdomain = parseSubdomain(jwtPayload.iss);
-    }
-    zoneId = jwtPayload.zid;
-  }
-
-  return new Promise((resolve, reject) => {
-    xssec.requests.requestClientCredentialsToken(
-      subdomain,
-      serviceCredentials,
-      undefined,
-      zoneId,
-      (err: Error, token: string, tokenResponse: ClientCredentialsResponse) =>
-        err ? reject(err) : resolve(tokenResponse)
-    );
-  });
-}
+import { replaceSubdomain } from './subdomain-replacer';
+import { refreshTokenGrant, userTokenGrant } from './legacy/xsuaa-service';
+import { UserTokenResponse } from './xsuaa-service-types';
+import { getClientCredentialsToken, getUserToken } from './xsuaa-service';
 
 /**
  * Returns an access token that can be used to call the given service. The token is fetched via a client credentials grant with the credentials of the given service.
@@ -91,7 +54,11 @@ export async function serviceToken(
   }
 
   try {
-    const token = await getClientCredentialsToken(service, options?.userJwt);
+    const token = await getClientCredentialsToken(
+      service,
+      options?.userJwt,
+      options
+    );
 
     if (opts.useCache) {
       clientCredentialsTokenCache.cacheRetrievedToken(
@@ -100,7 +67,7 @@ export async function serviceToken(
         token
       );
     }
-    // TODO: Token responses can have id_token instead of access_token (xssec does that correctly)
+
     return token.access_token;
   } catch (err) {
     throw new ErrorWithCause(
@@ -166,22 +133,14 @@ export async function jwtBearerToken(
   service: string | Service,
   options?: ResilienceOptions
 ): Promise<string> {
-  const xsuaa = multiTenantXsuaaCredentials(userJwt);
   const resolvedService = resolveService(service);
-  const serviceCreds = extractClientCredentials(resolvedService.credentials);
+
   const opts: ResilienceOptions = {
     enableCircuitBreaker: true,
     ...options
   };
 
-  return jwtBearerTokenGrant(xsuaa, serviceCreds, userJwt, opts)
-    .then((response: ClientCredentialsResponse) => response.access_token)
-    .catch(error => {
-      throw new ErrorWithCause(
-        `Fetching a JWT Bearer Token for service "${resolvedService.label}" failed!`,
-        error
-      );
-    });
+  return getUserToken(resolvedService, userJwt, opts);
 }
 
 function multiTenantXsuaaCredentials(userJwt?: string | JwtPayload) {